selesai buat fitur tambah detail transaksi, buat fitur tambah stock

+++ database/migrations/2024_03_11_035024_create_barang_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('barang', function (Blueprint $table) {
             $table->id();
             $table->string('nama_barang');
-            $table->integer('jumlah')->nullable();
+            $table->unsignedBigInteger('jumlah')->nullable();
             $table->integer('harga_modal');
             $table->timestamps();
         });
